refactor: ViewHolder 분리
