+++ app/src/main/java/com/dongyang/android/youdongknowme/ui/adapter/CafeteriaAnotherAdapter.kt
@@ -3,12 +3,11 @@ package com.dongyang.android.youdongknowme.ui.adapter
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.core.view.setPadding
 import androidx.recyclerview.widget.RecyclerView
-import com.dongyang.android.youdongknowme.R
 import com.dongyang.android.youdongknowme.databinding.ItemCafeteriaAnotherBinding
+import com.dongyang.android.youdongknowme.ui.viewholder.CafeteriaAnotherViewHolder
 
-class CafeteriaAnotherAdapter : RecyclerView.Adapter<CafeteriaAnotherAdapter.ViewHolder>() {
+class CafeteriaAnotherAdapter : RecyclerView.Adapter<CafeteriaAnotherViewHolder>() {
 
     init {
         setHasStableIds(true)
@@ -16,33 +15,8 @@ class CafeteriaAnotherAdapter : RecyclerView.Adapter<CafeteriaAnotherAdapter.Vie
 
     private var menu = arrayListOf<Array<String>>()
 
-    inner class ViewHolder(private val binding: ItemCafeteriaAnotherBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: Array<String>) {
-            binding.menuKr = item[0]
-            binding.menuEn = item[1].split('_')
-                .joinToString(" ") { word ->
-                    word.lowercase().replaceFirstChar { firstChar -> firstChar.uppercase() }
-                }
-            binding.menuPrice = item[2]
-
-            val context = binding.root.context
-            val resourceName = "img_cafeteria_" + item[1].lowercase()
-            val drawableResId =
-                context.resources.getIdentifier(resourceName, "drawable", context.packageName)
-
-            if (drawableResId != 0) {
-                binding.imgAnother.setImageResource(drawableResId)
-                binding.imgAnother.setPadding(0)
-            } else {
-                binding.imgAnother.setImageResource(R.drawable.img_cafeteria_korean)
-                binding.imgAnother.setPadding(15)
-            }
-        }
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        return ViewHolder(
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CafeteriaAnotherViewHolder {
+        return CafeteriaAnotherViewHolder(
             ItemCafeteriaAnotherBinding.inflate(
                 LayoutInflater.from(parent.context), parent, false
             )
@@ -53,7 +27,7 @@ class CafeteriaAnotherAdapter : RecyclerView.Adapter<CafeteriaAnotherAdapter.Vie
         return position.toLong()
     }
 
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: CafeteriaAnotherViewHolder, position: Int) {
         holder.bind(menu[position])
     }
